feat/add TS in the useForm, ForgotPassword, Login, Profile, Registration
