+++ src/hooks/useRefreshToken.js
@@ -4,8 +4,8 @@ import useSessionStorage from "./useSessionStorage";
 
 function useRefreshToken() {
   const { setCookie } = useCookies();
-  const { setToken } = useSessionStorage();
-  let token = JSON.parse(sessionStorage.getItem('refreshToken'));
+  const { setToken, getToken } = useSessionStorage();
+  let token = getToken('refreshToken');
 
   function postRefreshToken() {
     refreshToken(token).then(res => {
